LayoutTests/http/tests/inspector/target/pause-on-inline-debugger-statement-expected.txt

ALERT: PASS: Should pause on debugger statement.
Test page used to check that execution will break on inline 'debugger' statement after cross-origin navigation.
